+++ internal/multicluster/internal/types/computed_exported_services_test.go
@@ -1,9 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: BUSL-1.1
 
-//go:build consulent
-// +build consulent
-
 package types
 
 import (
@@ -17,40 +14,6 @@ import (
 	"github.com/stretchr/testify/require"
 	"testing"
 )
-
-func validComputedExportedServicesWithPeer() *multiclusterv1alpha1.ComputedExportedServices {
-	consumers := []*multiclusterv1alpha1.ComputedExportedService{
-		{
-			Consumers: []*multiclusterv1alpha1.ComputedExportedServicesConsumer{
-				{
-					ConsumerTenancy: &multiclusterv1alpha1.ComputedExportedServicesConsumer_Peer{
-						Peer: "peer",
-					},
-				},
-			},
-		},
-	}
-	return &multiclusterv1alpha1.ComputedExportedServices{
-		Consumers: consumers,
-	}
-}
-
-func validComputedExportedServicesWithPartition() *multiclusterv1alpha1.ComputedExportedServices {
-	consumers := []*multiclusterv1alpha1.ComputedExportedService{
-		{
-			Consumers: []*multiclusterv1alpha1.ComputedExportedServicesConsumer{
-				{
-					ConsumerTenancy: &multiclusterv1alpha1.ComputedExportedServicesConsumer_Partition{
-						Partition: "partition",
-					},
-				},
-			},
-		},
-	}
-	return &multiclusterv1alpha1.ComputedExportedServices{
-		Consumers: consumers,
-	}
-}
 
 func TestComputedExportedServicesValidations(t *testing.T) {
 	type testcase struct {
