+++ internal/multicluster/internal/types/namespace_exported_services_test.go
@@ -1,9 +1,6 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: BUSL-1.1
 
-//go:build consulent
-// +build consulent
-
 package types
 
 import (
@@ -17,70 +14,31 @@ import (
 	"testing"
 )
 
-func validPartitionExportedServicesWithPeer() *multiclusterv1alpha1.PartitionExportedServices {
-	consumers := []*multiclusterv1alpha1.ExportedServicesConsumer{
-		{
-			ConsumerTenancy: &multiclusterv1alpha1.ExportedServicesConsumer_Peer{
-				Peer: "peer",
-			},
-		},
-	}
-	return &multiclusterv1alpha1.PartitionExportedServices{
-		Consumers: consumers,
-	}
-}
-
-func validPartitionExportedServicesWithPartition() *multiclusterv1alpha1.PartitionExportedServices {
-	consumers := []*multiclusterv1alpha1.ExportedServicesConsumer{
-		{
-			ConsumerTenancy: &multiclusterv1alpha1.ExportedServicesConsumer_Partition{
-				Partition: "partition",
-			},
-		},
-	}
-	return &multiclusterv1alpha1.PartitionExportedServices{
-		Consumers: consumers,
-	}
-}
-
-func validPartitionExportedServicesWithSamenessGroup() *multiclusterv1alpha1.PartitionExportedServices {
-	consumers := []*multiclusterv1alpha1.ExportedServicesConsumer{
-		{
-			ConsumerTenancy: &multiclusterv1alpha1.ExportedServicesConsumer_SamenessGroup{
-				SamenessGroup: "sameness_group",
-			},
-		},
-	}
-	return &multiclusterv1alpha1.PartitionExportedServices{
-		Consumers: consumers,
-	}
-}
-
-func TestPartitionExportedServicesValidations(t *testing.T) {
+func TestNamespaceExportedServicesValidations(t *testing.T) {
 	type testcase struct {
 		Resource *pbresource.Resource
 	}
 	run := func(t *testing.T, tc testcase) {
-		err := ValidatePartitionExportedServices(tc.Resource)
+		err := ValidateNamespaceExportedServices(tc.Resource)
 		require.NoError(t, err)
 
-		resourcetest.MustDecode[*multiclusterv1alpha1.PartitionExportedServices](t, tc.Resource)
+		resourcetest.MustDecode[*multiclusterv1alpha1.NamespaceExportedServices](t, tc.Resource)
 	}
 
 	cases := map[string]testcase{
-		"partition exported services with peer": {
-			Resource: resourcetest.Resource(multiclusterv1alpha1.PartitionExportedServicesType, "partition-exported-services").
-				WithData(t, validPartitionExportedServicesWithPeer()).
+		"namespace exported services with peer": {
+			Resource: resourcetest.Resource(multiclusterv1alpha1.NamespaceExportedServicesType, "namespace-exported-services").
+				WithData(t, validNamespaceExportedServicesWithPeer()).
 				Build(),
 		},
-		"partition exported services with partition": {
-			Resource: resourcetest.Resource(multiclusterv1alpha1.PartitionExportedServicesType, "partition-exported-services").
-				WithData(t, validPartitionExportedServicesWithPartition()).
+		"namespace exported services with partition": {
+			Resource: resourcetest.Resource(multiclusterv1alpha1.NamespaceExportedServicesType, "namespace-exported-services").
+				WithData(t, validNamespaceExportedServicesWithPartition()).
 				Build(),
 		},
-		"partition exported services with sameness_group": {
-			Resource: resourcetest.Resource(multiclusterv1alpha1.PartitionExportedServicesType, "partition-exported-services").
-				WithData(t, validPartitionExportedServicesWithSamenessGroup()).
+		"namespace exported services with sameness_group": {
+			Resource: resourcetest.Resource(multiclusterv1alpha1.NamespaceExportedServicesType, "namespace-exported-services").
+				WithData(t, validNamespaceExportedServicesWithSamenessGroup()).
 				Build(),
 		},
 	}
@@ -92,7 +50,7 @@ func TestPartitionExportedServicesValidations(t *testing.T) {
 	}
 }
 
-func TestPartitionExportedServicesACLs(t *testing.T) {
+func TestNamespaceExportedServicesACLs(t *testing.T) {
 	// Wire up a registry to generically invoke hooks
 	registry := resource.NewRegistry()
 	Register(registry)
@@ -128,12 +86,12 @@ func TestPartitionExportedServicesACLs(t *testing.T) {
 		}
 	}
 
-	reg, ok := registry.Resolve(multiclusterv1alpha1.PartitionExportedServicesType)
+	reg, ok := registry.Resolve(multiclusterv1alpha1.NamespaceExportedServicesType)
 	require.True(t, ok)
 
 	run := func(t *testing.T, tc testcase) {
-		exportedServiceData := &multiclusterv1alpha1.PartitionExportedServices{}
-		res := resourcetest.Resource(multiclusterv1alpha1.PartitionExportedServicesType, "partition-exported-services").
+		exportedServiceData := &multiclusterv1alpha1.NamespaceExportedServices{}
+		res := resourcetest.Resource(multiclusterv1alpha1.NamespaceExportedServicesType, "namespace-exported-services").
 			WithData(t, exportedServiceData).
 			Build()
 		resourcetest.ValidateAndNormalize(t, registry, res)
